# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFtaXJleDEyOEBnbWFpbC5jb20iLCJleHAiOjUyNjgzMTg5OTYsImV4cGlyZV9hdCI6IiIsImZpcnN0bmFtZSI6Itin2YXbjNixIiwiaWQiOjEsImxhc3RuYW1lIjoi2LTbjNix2K_ZhNuMIiwibW9iaWxlIjoiMDkwMjQ4MDk3NTAiLCJvcmlnX2lhdCI6MTY2ODMyMjU5Niwic3RhdHVzIjoiIn0.x7BKuxw288cm1JsskGRD178UPmNz-xRwkWHtb0WsU74",
        "title": "Selloora Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://www.amirshirdel.ir",
            "email": "amirex128@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8585",
    "basePath": "/",
    "paths": {
...
        "/user/address/create": {
            "post": {
                "description": "کاربران میتوانند برای خود لیستی از ادرس های مختلف ایجاد کنند تا هر بار به راحتی مشخص کنند محصول خود را میخواند از کدام ادرس ارسال نمایید",
                "tags": [
                    "address"
                ],
                "summary": "ایجاد آدرس جدید",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.CreateAddress"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/address/delete/{id}": {
            "post": {
                "description": "کاربران میتوانند برای خود لیستی از ادرس های مختلف ایجاد کنند تا هر بار به راحتی مشخص کنند محصول خود را میخواند از کدام ادرس ارسال نمایید",
                "tags": [
                    "address"
                ],
                "summary": "حذف آدرس",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه آدرس",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/address/list": {
            "get": {
                "description": "کاربران میتوانند برای خود لیستی از ادرس های مختلف ایجاد کنند تا هر بار به راحتی مشخص کنند محصول خود را میخواند از کدام ادرس ارسال نمایید",
                "tags": [
                    "address"
                ],
                "summary": "لیست آدرس",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "متن جستجو",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "شماره صفحه",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "تعداد صفحه",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "مرتب سازی براساس desc/asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/user/address/update": {
            "post": {
                "description": "کاربران میتوانند برای خود لیستی از ادرس های مختلف ایجاد کنند تا هر بار به راحتی مشخص کنند محصول خود را میخواند از کدام ادرس ارسال نمایید",
                "tags": [
                    "address"
                ],
                "summary": "ویرایش آدرس",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.UpdateAddress"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/category": {
            "get": {
                "description": "کاربران برای دسته بندی کردن محصولات خود و مقالات خود از این دسته بندی ها استفاده میکنند که دو نوع میباشد نوع اول برای محصولات و نوع دوم ان برای مقالات این دو نوع از هم جدا هستن ولی از یک ای پی ای ساخته می شوند و نمایش داده میشوند",
                "tags": [
                    "category"
                ],
                "summary": "لیست دسته بندی ها",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "متن جستجو",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "شماره صفحه",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "تعداد صفحه",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "مرتب سازی براساس desc/asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/user/category/create": {
            "post": {
                "description": "کاربران برای دسته بندی کردن محصولات خود و مقالات خود از این دسته بندی ها استفاده میکنند که دو نوع میباشد نوع اول برای محصولات و نوع دوم ان برای مقالات این دو نوع از هم جدا هستن ولی از یک ای پی ای ساخته می شوند و نمایش داده میشوند",
                "tags": [
                    "category"
                ],
                "summary": "ایجاد دسته بندی",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.CreateCategory"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/category/delete/{id}": {
            "post": {
                "description": "کاربران برای دسته بندی کردن محصولات خود و مقالات خود از این دسته بندی ها استفاده میکنند که دو نوع میباشد نوع اول برای محصولات و نوع دوم ان برای مقالات این دو نوع از هم جدا هستن ولی از یک ای پی ای ساخته می شوند و نمایش داده میشوند",
                "tags": [
                    "category"
                ],
                "summary": "حذف دسته بندی",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه آدرس",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/category/update": {
            "post": {
                "description": "کاربران برای دسته بندی کردن محصولات خود و مقالات خود از این دسته بندی ها استفاده میکنند که دو نوع میباشد نوع اول برای محصولات و نوع دوم ان برای مقالات این دو نوع از هم جدا هستن ولی از یک ای پی ای ساخته می شوند و نمایش داده میشوند",
                "tags": [
                    "category"
                ],
                "summary": "ویرایش دسته بندی ها",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.UpdateCategory"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "DTOs.CreateAddress": {
            "type": "object",
            "required": [
                "address",
                "city_id",
                "full_name",
                "mobile",
                "postal_code",
                "province_id",
                "title"
            ],
            "properties": {
                "address": {
                    "description": "ادرس کامل",
                    "type": "string",
                    "example": "ادرس کامل"
                },
                "city_id": {
                    "description": "شناسه شهر",
                    "type": "integer",
                    "example": 1
                },
                "full_name": {
                    "description": "نام گیرنده",
                    "type": "string",
                    "example": "نام گیرنده"
                },
                "lat": {
                    "description": "عرض جغذافیایی",
                    "type": "string",
                    "example": ""
                },
                "long": {
                    "description": "طول جقرافیایی",
                    "type": "string",
                    "example": ""
                },
                "mobile": {
                    "description": "موبایل",
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 11,
                    "example": "09024809750"
                },
                "postal_code": {
                    "description": "کد پستی",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 10,
                    "example": "1111111111"
                },
                "province_id": {
                    "description": "شناسه استان",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "عنوان",
                    "type": "string",
                    "example": "عنوان"
                }
            }
        },
        "DTOs.CreateCategory": {
            "type": "object",
            "required": [
                "name",
                "parent_id",
                "type"
            ],
            "properties": {
                "description": {
                    "description": "توضیحات دسته بندی",
                    "type": "string",
                    "example": "توضیحات دسته بندی"
                },
                "equivalent": {
                    "description": "کلمات مترادف که با کاما از هم جدا شده اند",
                    "type": "string",
                    "example": "آموزش,یادگیری"
                },
                "gallery_id": {
                    "description": "شناسه تصویر آپلود شده",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "نام دسته بندی",
                    "type": "string",
                    "example": "نام دسته بندی"
                },
                "parent_id": {
                    "description": "شناسه دسته بندی پدر",
                    "type": "integer",
                    "example": 0
                },
                "type": {
                    "description": "نوع دسته بندی",
                    "type": "string",
                    "enum": [
                        "product",
                        "post"
                    ],
                    "example": "product"
                }
            }
        },
...
        "DTOs.UpdateAddress": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "address": {
                    "description": "ادرس کامل",
                    "type": "string",
                    "example": "ادرس کامل"
                },
                "city_id": {
                    "description": "شناسه شهر",
                    "type": "integer",
                    "example": 1
                },
                "full_name": {
                    "description": "نام گیرنده",
                    "type": "string",
                    "example": "نام گیرنده"
                },
                "id": {
                    "description": "شناسه آدرس",
                    "type": "integer",
                    "example": 1
                },
                "lat": {
                    "description": "عرض جغذافیایی",
                    "type": "string",
                    "example": "35.123456"
                },
                "long": {
                    "description": "طول جقرافیایی",
                    "type": "string",
                    "example": "35.123456"
                },
                "mobile": {
                    "description": "موبایل",
                    "type": "string",
                    "example": "09024809750"
                },
                "postal_code": {
                    "description": "کد پستی",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 10,
                    "example": "1111111111"
                },
                "province_id": {
                    "description": "شناسه استان",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "عنوان",
                    "type": "string",
                    "example": "عنوان"
                }
            }
        },
        "DTOs.UpdateCategory": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "description": "توضیحات دسته بندی",
                    "type": "string",
                    "example": "توضیحات دسته بندی"
                },
                "equivalent": {
                    "description": "کلمات مترادف که با کاما از هم جدا شده اند",
                    "type": "string",
                    "example": "آموزش,یادگیری"
                },
                "gallery_id": {
                    "description": "شناسه تصویر آپلود شده",
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "description": "شناسه دسته بندی برای ویرایش",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "نام دسته بندی",
                    "type": "string",
                    "example": "نام دسته بندی"
                },
                "parent_id": {
                    "description": "شناسه دسته بندی پدر",
                    "type": "integer",
                    "example": 0
                },
                "sort": {
                    "description": "ترتیب مرتب سازی دسته بندی",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}